Change to class and get accordions working
Also made some file sorting changes
