+++ components/AccordionFiles/entertainmentAccordion.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { View, Text, StyleSheet } from "react-native";
-import EverydayValues from "../data/everyday.json";
+import EverydayValues from "../../data/everyday.json";
 
 export default function HappyHourAccordion(name) {
   return (
