+++ components/picture_linker.js
@@ -2,6 +2,7 @@ import React from "react";
 import barImages from "../assets/bar_list.js";
 
 function getBarLink(bar) {
+  //console.log(bar);
   if (bar == "Champs") {
     return barImages.Champs;
   } else if (bar == "Phyrst") {
